requirements.txt
click==6.7
Flask==0.12.2
Flask-Views==0.2.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
google-cloud-datastore
google-cloud-vision
google-cloud-storage
google-cloud-translate
gunicorn
